+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Unreleased
 
+  * Add DTLS 1.2 ChaCha20 and X25519 support #77
   * Bump MSRV to 1.85.0 #75
   * Make cipher and kx configurable #73